+++ CHANGELOG.md
@@ -24,7 +24,7 @@ Test/interim releases are not documented here.
 
 ---
 
-## WIP
+## [0.3.3]
 ### Changed
 - Transpiler - only transpile interfaces that have IFinObj as an ancestor
     - https://github.com/fin-language/fin/issues/63